+++ cognite/client/_api/diagrams.py
@@ -29,9 +29,9 @@ class DiagramsAPI(APIClient):
 
     def __init__(self, config: ClientConfig, api_version: Optional[str], cognite_client: CogniteClient) -> None:
         super().__init__(config, api_version, cognite_client)
-        # https://docs.cognite.com/api/playground/#tag/Engineering-diagrams/operation/diagramDetect
+        # https://docs.cognite.com/api/playground/#operation/diagramDetect
         self._DETECT_API_FILE_LIMIT = 50
-        # https://docs.cognite.com/api/playground/#tag/Engineering-diagrams/operation/diagramDetectMultipleResults
+        # https://docs.cognite.com/api/playground/#operation/diagramDetectMultipleResults
         self._DETECT_API_STATUS_JOB_LIMIT = 1000
 
     def _camel_post(